requirements.txt
Pillow==9.5.0
qrcode
streamlit==1.24.1
pyshorteners==1.0.1